Don't try to fetch the Person for the current user if not authenticated
We try to special case the Person object matching the current user since it is used
quite often. Since we now are more strict with fetch user data for unauthenticated
requests we need to skip this if the user is not authenticated.

Also we need to ignore the current user if it is a service account, since those don't
exist in LDAP.